ssl: add a note about PKCS#12 password
For some reason, the key is protected inside the keystore by the
password from the PKCS#12 file. We should be able to chnage it with
`-destkeypass` but it said that we should have aliases and when we use
`-alias`, it complains about conflicts.

Just tell the user to ensure the PKCS#12 have the same password.