prefs: add "2 minutes" and "5 minutes" as possible timeouts
On some JS-heavy page, 1 minute is sometimes too short. As the global
problem, JS is single-threaded and therefore heavy computation may block
the same thread that should execute the timeout function. We could use a
worker to fix that but this would defeat our ability to detect a page is
stuck.